+++ .dbInfo
@@ -1,0 +1,3 @@
+USERNAME="to_user"
+PASSWORD="twelve"
+ENVIRONMENT="to_development"
