--- newAsn
@@ -1,1 +1,0 @@
-[{"asn":21508,"cachegroup":15,"id":1,"last_updated":"2012-09-17 15:41:22"}]
