--- newView
@@ -1,2 +1,0 @@
-[{"Name":"newView", 
-"Query":"select id from asn"}]
