--- newViews
@@ -1,4 +1,0 @@
-[{"Name":"newView", 
-  "Query":"select id from asn"},
- {"Name":"newView2",
-  "Query":"select id, cachegroup from asn"}]
