+++ testFiles/newAsn
@@ -1,0 +1,1 @@
+[{"asn":21508,"cachegroup":15,"last_updated":"2012-09-17 15:41:22"}]
