+++ testFiles/newAsns
@@ -1,0 +1,1 @@
+[{"asn":21508,"cachegroup":12,"id":1001,"last_updated":"2012-09-17 15:41:22"},{"asn":21508,"cachegroup":15,"id":1002,"last_updated":"2012-09-17 15:41:22"}]
