+++ testFiles/newView2
@@ -1,0 +1,2 @@
+[{"Name":"newView", 
+"Query":"select id from asn"}]
