+++ testFiles/newViews
@@ -1,0 +1,4 @@
+[{"Name":"newView", 
+  "Query":"select id from asn"},
+ {"Name":"newView2",
+  "Query":"select id, cachegroup from asn"}]
